+++ src/main/java/hanium/modic/backend/domain/user/enums/UserConstant.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.domain.user.entity;
+package hanium.modic.backend.domain.user.enums;
 
 import lombok.Getter;
